* Fixed LG_REFLECTSHIELD & SC_SHADOWFORM. Royal Guards under Reflect damage are reflecting damage received from shadow form.
git-svn-id: https://svn.code.sf.net/p/v1-3ceam/code/trunk@387 cad27aaa-dce3-4a30-a00a-e4fd67c11881

+++ rewrite/src/map/clif.c
@@ -4151,7 +4151,7 @@ int clif_damage(struct block_list* src, struct block_list* dst, unsigned int tic
 		WBUFW(buf,22) = min(damage, SHRT_MAX);
 		WBUFW(buf,i + 27) = damage2;
 #else
-		WBUFL(buf,22) = min(damage, SHRT_MAX);
+		WBUFL(buf,22) = min(damage, INT_MAX);
 		WBUFL(buf,i+ 27) = damage2;
 #endif
 	}
